backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["David Saidov <delightxxls@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
djoser = "^2.2.2"
psycopg = {extras = ["binary"], version = "^3.1.19"}
django-cors-headers = "^4.4.0"
celery = "^5.4.0"
django-redis = "^5.4.0"
eventlet = "^0.36.1"
pillow = "^10.4.0"
django-phonenumber-field = "^8.0.0"
django-phonenumbers = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
